# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bank Rest API",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/accounts": {
      "get": {
        "tags": ["Account"],
        "summary": "Get accounts",
        "description": "Get accounts",
        "operationId": "getAccounts",
        "responses": {
          "200": {
            "description": "200 - List accounts",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "492"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 25 Oct 2024 12:57:06 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"1ec-qJy7rIuhelpYtp4gJ7LRH5im0bw\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "balance": {
                            "type": "string",
                            "example": "0"
                          },
                          "bankAccountNumber": {
                            "type": "string",
                            "example": "1"
                          },
                          "bankName": {
                            "type": "string",
                            "example": "BCA"
                          },
                          "createdAt": {
                            "type": "string",
                            "example": "2024-10-25T12:54:29.870Z"
                          },
                          "id": {
                            "type": "string",
                            "example": "6db74fb7-6243-4f7e-a645-4d2336ecbb17"
                          },
                          "updatedAt": {
                            "type": "string",
                            "example": "2024-10-25T12:54:29.870Z"
                          },
                          "user": {
                            "type": "object",
                            "properties": {
                              "createdAt": {
                                "type": "string",
                                "example": "2024-10-25T12:51:28.774Z"
                              },
                              "email": {
                                "type": "string",
                                "example": "aminrisal@gmail.com"
                              },
                              "id": {
                                "type": "string",
                                "example": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                              },
                              "name": {
                                "type": "string",
                                "example": "Risal Amin"
                              },
                              "password": {
                                "type": "string",
                                "example": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq"
                              },
                              "updatedAt": {
                                "type": "string",
                                "example": "2024-10-25T12:51:28.774Z"
                              }
                            }
                          },
                          "userId": {
                            "type": "string",
                            "example": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                          }
                        }
                      },
                      "example": [
                        {
                          "balance": "0",
                          "bankAccountNumber": "1",
                          "bankName": "BCA",
                          "createdAt": "2024-10-25T12:54:29.870Z",
                          "id": "6db74fb7-6243-4f7e-a645-4d2336ecbb17",
                          "updatedAt": "2024-10-25T12:54:29.870Z",
                          "user": {
                            "createdAt": "2024-10-25T12:51:28.774Z",
                            "email": "aminrisal@gmail.com",
                            "id": "5ec59a42-64f4-437d-8477-3b3db84c60aa",
                            "name": "Risal Amin",
                            "password": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq",
                            "updatedAt": "2024-10-25T12:51:28.774Z"
                          },
                          "userId": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                        }
                      ]
                    }
                  }
                },
                "examples": {
                  "200 - List accounts": {
                    "value": {
                      "data": [
                        {
                          "balance": "0",
                          "bankAccountNumber": "1",
                          "bankName": "BCA",
                          "createdAt": "2024-10-25T12:54:29.870Z",
                          "id": "6db74fb7-6243-4f7e-a645-4d2336ecbb17",
                          "updatedAt": "2024-10-25T12:54:29.870Z",
                          "user": {
                            "createdAt": "2024-10-25T12:51:28.774Z",
                            "email": "aminrisal@gmail.com",
                            "id": "5ec59a42-64f4-437d-8477-3b3db84c60aa",
                            "name": "Risal Amin",
                            "password": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq",
                            "updatedAt": "2024-10-25T12:51:28.774Z"
                          },
                          "userId": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Account"],
        "summary": "Create account",
        "description": "Create account",
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bankName": {
                    "type": "string",
                    "example": "BCA"
                  },
                  "userId": {
                    "type": "string",
                    "example": "5ec59a42-64f4-437d-8477-3b3db84c60ab"
                  }
                }
              },
              "examples": {
                "Create account": {
                  "value": {
                    "bankName": "BCA",
                    "userId": "5ec59a42-64f4-437d-8477-3b3db84c60ab"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "201 - Create account",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "490"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Mon, 28 Oct 2024 16:04:42 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"1ea-CONu7djMopuSKw1KSb4HcbIVtHQ\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "balance": {
                          "type": "string",
                          "example": "0"
                        },
                        "bankAccountNumber": {
                          "type": "string",
                          "example": "7"
                        },
                        "bankName": {
                          "type": "string",
                          "example": "BRI"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2024-10-28T16:04:42.759Z"
                        },
                        "id": {
                          "type": "string",
                          "example": "dafcf049-2056-4a1e-8674-0cd9f9a15369"
                        },
                        "updatedAt": {
                          "type": "string",
                          "example": "2024-10-28T16:04:42.759Z"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "createdAt": {
                              "type": "string",
                              "example": "2024-10-25T12:51:28.774Z"
                            },
                            "email": {
                              "type": "string",
                              "example": "aminrisal@gmail.com"
                            },
                            "id": {
                              "type": "string",
                              "example": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                            },
                            "name": {
                              "type": "string",
                              "example": "Risal Amin"
                            },
                            "password": {
                              "type": "string",
                              "example": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq"
                            },
                            "updatedAt": {
                              "type": "string",
                              "example": "2024-10-25T12:51:28.774Z"
                            }
                          }
                        },
                        "userId": {
                          "type": "string",
                          "example": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "201 - Create account": {
                    "value": {
                      "data": {
                        "balance": "0",
                        "bankAccountNumber": "7",
                        "bankName": "BRI",
                        "createdAt": "2024-10-28T16:04:42.759Z",
                        "id": "dafcf049-2056-4a1e-8674-0cd9f9a15369",
                        "updatedAt": "2024-10-28T16:04:42.759Z",
                        "user": {
                          "createdAt": "2024-10-25T12:51:28.774Z",
                          "email": "aminrisal@gmail.com",
                          "id": "5ec59a42-64f4-437d-8477-3b3db84c60aa",
                          "name": "Risal Amin",
                          "password": "$2b$10$NQZCR6olPgVJXaqceA8s.eL9zFGyR5SCR/0z1pTDvcbiqcLcQaIFq",
                          "updatedAt": "2024-10-25T12:51:28.774Z"
                        },
                        "userId": "5ec59a42-64f4-437d-8477-3b3db84c60aa"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "400 - Invalid body",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "63"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 25 Oct 2024 12:56:41 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"3f-9VWshTA/wj70REP2S7mArRQDKb0\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "\"bankName\" is not allowed to be empty"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "400 - Invalid body": {
                    "value": {
                      "error": {
                        "message": "\"bankName\" is not allowed to be empty"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "401 - Invalid token",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "37"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Mon, 28 Oct 2024 16:05:40 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"25-FIV/ZCaZ465tIhBUa/pSs9dwkFc\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Invalid token"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "401 - Invalid token": {
                    "value": {
                      "error": {
                        "message": "Invalid token"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "404 - User not found",
            "headers": {
              "Access-Control-Allow-Credentials": {
                "schema": {
                  "type": "string",
                  "example": "true"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "http://localhost:3000"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "38"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 25 Oct 2024 12:56:09 GMT"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"26-BsgGbTj6n3ZmvCCse0wan/1YC1c\""
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              },
              "Vary": {
                "schema": {
                  "type": "string",
                  "example": "Origin"
                }
              },
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "User not found"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "404 - User not found": {
                    "value": {
                      "error": {
                        "message": "User not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}